@@ -3,9 +3,8 @@ publish: none
 description: xlider for elementary
 version: 1.0.0
 
-
 environment:
-  sdk: '>=2.18.1 <3.0.0'
+  sdk: ">=3.0.2 <4.0.0"
 
 dependencies:
   flutter:
@@ -14,6 +13,6 @@ dependencies:
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.1
+  coolint: 1.0.2
 
 flutter:
